@@ -1,22 +1,23 @@
 name: weather_forecast
 description: A sample Flutter application by Seyed Abbas Ghomi.
 
-publish_to: 'none'
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 0.1.0+1
+version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  cupertino_icons: ^1.0.0
-  mobx: ^1.2.1+4
-  http: ^0.12.2
-  sailor: ^0.7.1
-  provider: ^4.3.2+2
-  equatable: ^1.2.5
-  google_maps_flutter: ^1.0.5
-  flutter_staggered_grid_view: ^0.3.2
+  cupertino_icons: ^1.0.3
+  mobx: ^2.0.1
+  http: ^0.13.3
+  beamer: ^0.13.3
+  provider: ^5.0.0
+  equatable: ^2.0.2
+  flutter_mobx: ^2.0.0
+  google_maps_flutter: ^2.0.6
+  flutter_staggered_grid_view: ^0.4.0
 
   flutter:
     sdk: flutter
@@ -24,10 +25,10 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.10.4
-  mobx_codegen: ^1.1.1+3
+  build_runner: ^2.0.4
+  mobx_codegen: ^2.0.1+3
+  pedantic: ^1.11.0
 
 flutter:
 
   uses-material-design: true
-
