@@ -1,4 +1,3 @@
-
 import 'package:mobx/mobx.dart';
 import 'package:weather_forecast/common/api_functions/weather/get_current_weather_data_api.dart';
 import 'package:weather_forecast/common/constants/enums.dart';
@@ -7,39 +6,43 @@ import 'package:weather_forecast/common/models/weather/weather_response_model.da
 
 part 'weather_store.g.dart';
 
-class WeatherStore extends _WeatherStore with _$WeatherStore {
-}
+class WeatherStore extends _WeatherStore with _$WeatherStore {}
 
 abstract class _WeatherStore with Store {
+  @observable
+  ObservableFuture<WeatherResponseModel?>? _responseFuture;
 
   @observable
-  ObservableFuture<WeatherResponseModel> _responseFuture;
-  @observable
-  WeatherResponseModel response;
+  WeatherResponseModel? response;
 
   @observable
-  String error;
+  String error = '';
 
   @computed
   StoreState get state {
-    if (_responseFuture == null ||
-        _responseFuture.status == FutureStatus.rejected) {
+    if (response == null) {
       return StoreState.initial;
     }
-    return _responseFuture.status == FutureStatus.pending
-        ? StoreState.service_called
-        :  response==null? StoreState.error :  StoreState.data_received;
+    if (_responseFuture!.status == FutureStatus.rejected) {
+      return StoreState.initial;
+    }
+    if (_responseFuture!.status == FutureStatus.pending) {
+      return StoreState.service_called;
+    }
+    return StoreState.data_received;
   }
 
   @action
   Future getWeather(String cityName) async {
+    if( cityName.isEmpty){
+      return;
+    }
     try {
-      error = null;
-      _responseFuture =
-          ObservableFuture( getCurrentWeatherDataApi(new WeatherRequestModel(cityName: cityName)));
+      error = '';
+      _responseFuture = ObservableFuture(getCurrentWeatherDataApi(WeatherRequestModel(cityName: cityName)));
       response = await _responseFuture;
     } on Exception {
-      error = "Fault in fetching weather data, Check network your phone connection";
+      error = 'Fault in fetching weather data, Check network your phone connection';
     }
   }
 }
