@@ -1,43 +1,50 @@
 import 'dart:async';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:mobx/mobx.dart';
 import 'package:provider/provider.dart';
 import 'package:weather_forecast/common/constants/enums.dart';
-import 'package:weather_forecast/common/mobx/flutter_mobx.dart';
+import 'package:weather_forecast/common/models/weather/weather_response_model.dart';
 import 'package:weather_forecast/common/widgets/app_bar_widget.dart';
 import 'package:weather_forecast/common/widgets/grid_widget.dart';
 import 'package:weather_forecast/common/widgets/main_menu_widget.dart';
 import 'package:weather_forecast/common/stores/weather/weather_store.dart';
 
 class HomePage extends StatefulWidget {
-  HomePage({Key key, this.title}) : super(key: key);
-  final String title;
-
   @override
   _HomePageState createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
-  GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey();
-  List<ReactionDisposer> _reactionDisposersList;
+  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey();
+  late List<ReactionDisposer> _reactionDisposersList;
 
   //Text widgets titles
   final String idleTextTitle = 'Listening to your thoughts...';
   final String listeningTextTitle = 'Nothing happening...';
 
   //to check if user is typing or not debounce timer is used
-  Timer _debounceTimer;
+  late Timer _debounceTimer;
 
   //Mobx store for weather
-  WeatherStore _weatherStore;
+  late WeatherStore _weatherStore;
 
   //City name text field controller
-  var _cityNameTextEditingController = TextEditingController();
+  final _cityNameTextEditingController = TextEditingController();
+
+  late Size screenSize;
+
+  WeatherResponseModel? response;
 
   @override
   void initState() {
     super.initState();
+    _debounceTimer = Timer(const Duration(milliseconds: 500), () {
+      _weatherStore.getWeather(_cityNameTextEditingController.text);
+    });
+    _debounceTimer.cancel();
     _cityNameTextEditingController.addListener(_onCityNameChanged);
   }
 
@@ -51,19 +58,17 @@ class _HomePageState extends State<HomePage> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    _weatherStore ??= Provider.of<WeatherStore>(context);
-    _reactionDisposersList ??= [
-      reaction((_) => _weatherStore.error, (String message) {
-        _scaffoldKey.currentState.showSnackBar(
+    _weatherStore = Provider.of<WeatherStore>(context);
+    _reactionDisposersList = [
+      reaction((_) => _weatherStore.error, (message) {
+        ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text(message),
+            content: Text(message.toString()),
           ),
         );
       }),
     ];
   }
-
-  Size screenSize;
 
   @override
   Widget build(BuildContext context) {
@@ -71,66 +76,64 @@ class _HomePageState extends State<HomePage> {
     _weatherStore = Provider.of<WeatherStore>(context);
     return Container(
       decoration: BoxDecoration(
-          gradient: LinearGradient(
-              begin: Alignment.topRight,
-              end: Alignment.bottomLeft,
-              stops: [
-            -0.7,
-            1.7,
-          ],
-              colors: [
-            Colors.yellow,
-            Colors.red,
-          ])),
+          gradient: LinearGradient(begin: Alignment.topRight, end: Alignment.bottomLeft, stops: [
+        -0.7,
+        1.7,
+      ], colors: [
+        Colors.yellow,
+        Colors.red,
+      ])),
       child: Scaffold(
         key: _scaffoldKey,
         backgroundColor: Colors.transparent,
         appBar: AppBarWidget(scaffoldKey: _scaffoldKey),
         extendBodyBehindAppBar: true,
         drawer: MainMenuWidget(scaffoldKey: _scaffoldKey),
-        body: Container(
-          width: screenSize.width,
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Text(
-                'City name',
-                textAlign: TextAlign.center,
+        body:
+              Column(
+                mainAxisSize: MainAxisSize.max,
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  Container(height: 10,),
+                  Text(
+                    'City name',
+                    textAlign: TextAlign.center,
+                  ),
+                  Container(
+                    width: screenSize.width * 70 / 100,
+                    child: TextField(
+                      controller: _cityNameTextEditingController,
+                    ),
+                  ),
+                   SingleChildScrollView(
+                      child: Observer(builder: (context) {
+                        response = _weatherStore.response;
+                        return checkWeatherState(response);
+                      }),
+                    ),
+
+                ],
               ),
-              Container(
-                width: screenSize.width * 70 / 100,
-                child: TextField(
-                  controller: _cityNameTextEditingController,
-                ),
-              ),
-              SingleChildScrollView(
-                child: Observer(builder: (context) {
-                  return checkWeatherState();
-                }),
-              ),
-            ],
-          ),
-        ),
       ),
     );
   }
 
-  _onCityNameChanged() {
-    if (_debounceTimer?.isActive ?? false) _debounceTimer.cancel();
-    _debounceTimer = Timer(const Duration(milliseconds: 500), () {
-      _weatherStore.getWeather(_cityNameTextEditingController.text);
+  void _onCityNameChanged() {
+    if (_debounceTimer.isActive) _debounceTimer.cancel();
+    _debounceTimer = Timer(const Duration(milliseconds: 500), () async {
+      await _weatherStore.getWeather(_cityNameTextEditingController.text);
     });
   }
 
-  Widget checkWeatherState() {
+  Widget checkWeatherState(WeatherResponseModel? response) {
     switch (_weatherStore.state) {
       case StoreState.initial:
         return Text(idleTextTitle, textAlign: TextAlign.center);
       case StoreState.service_called:
         return Text(idleTextTitle, textAlign: TextAlign.center);
       case StoreState.data_received:
-        return GridWidget(
-          weatherResponseModel: _weatherStore.response,
+        return  GridWidget(
+          weatherResponseModel: response,
         );
       case StoreState.error:
         return Text(idleTextTitle, textAlign: TextAlign.center);
