@@ -2,17 +2,17 @@
 class Clouds {
     int all;
 
-    Clouds({this.all});
+    Clouds({required this.all});
 
     factory Clouds.fromJson(Map<String, dynamic> json) {
         return Clouds(
-            all: json['all'], 
+            all: json['all'] as int,
         );
     }
 
     Map<String, dynamic> toJson() {
-        final Map<String, dynamic> data = new Map<String, dynamic>();
-        data['all'] = this.all;
+        final data = <String, dynamic>{};
+        data['all'] = all;
         return data;
     }
 }
