@@ -1,27 +1,26 @@
-
 class Weather {
-    String description;
-    String icon;
-    int id;
-    String main;
+  String description;
+  String icon;
+  int id;
+  String main;
 
-    Weather({this.description, this.icon, this.id, this.main});
+  Weather({required this.description, required this.icon, required this.id, required this.main});
 
-    factory Weather.fromJson(Map<String, dynamic> json) {
-        return Weather(
-            description: json['description'], 
-            icon: json['icon'], 
-            id: json['id'], 
-            main: json['main'], 
-        );
-    }
+  factory Weather.fromJson(Map<String, dynamic> json) {
+    return Weather(
+      description: json['description'] as String,
+      icon: json['icon'] as String,
+      id: json['id'] as int,
+      main: json['main'] as String,
+    );
+  }
 
-    Map<String, dynamic> toJson() {
-        final Map<String, dynamic> data = new Map<String, dynamic>();
-        data['description'] = this.description;
-        data['icon'] = this.icon;
-        data['id'] = this.id;
-        data['main'] = this.main;
-        return data;
-    }
+  Map<String, dynamic> toJson() {
+    var data = <String, dynamic>{};
+    data['description'] = description;
+    data['icon'] = icon;
+    data['id'] = id;
+    data['main'] = main;
+    return data;
+  }
 }
