@@ -1,20 +1,19 @@
-
 class WeatherRequestModel {
   final String cityName;
 
   WeatherRequestModel({
-    this.cityName,
+    required this.cityName,
   });
 
   factory WeatherRequestModel.fromJson(Map<String, dynamic> json) {
     return WeatherRequestModel(
-      cityName: json['cityName'],
+      cityName: json['cityName'] as String,
     );
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['cityName'] = this.cityName;
+    var data = <String, dynamic>{};
+    data['cityName'] = cityName;
     return data;
   }
 }
