@@ -2,6 +2,7 @@
 import 'package:equatable/equatable.dart';
 
 abstract class DrawerStoreState extends Equatable{
+  @override
   List<Object> get props =>[];
 }
 
@@ -9,7 +10,8 @@ class NewLocationOnMapState extends DrawerStoreState {
   final double latitude;
   final double longitude;
 
-  NewLocationOnMapState({this.latitude, this.longitude});
+  NewLocationOnMapState({required this.latitude, required this.longitude});
 
+  @override
   List<Object> get props => [latitude, longitude];
 }
