@@ -9,7 +9,7 @@ part of 'weather_store.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$WeatherStore on _WeatherStore, Store {
-  Computed<StoreState> _$stateComputed;
+  Computed<StoreState>? _$stateComputed;
 
   @override
   StoreState get state => (_$stateComputed ??=
@@ -19,13 +19,13 @@ mixin _$WeatherStore on _WeatherStore, Store {
   final _$_responseFutureAtom = Atom(name: '_WeatherStore._responseFuture');
 
   @override
-  ObservableFuture<WeatherResponseModel> get _responseFuture {
+  ObservableFuture<WeatherResponseModel?>? get _responseFuture {
     _$_responseFutureAtom.reportRead();
     return super._responseFuture;
   }
 
   @override
-  set _responseFuture(ObservableFuture<WeatherResponseModel> value) {
+  set _responseFuture(ObservableFuture<WeatherResponseModel?>? value) {
     _$_responseFutureAtom.reportWrite(value, super._responseFuture, () {
       super._responseFuture = value;
     });
@@ -34,13 +34,13 @@ mixin _$WeatherStore on _WeatherStore, Store {
   final _$responseAtom = Atom(name: '_WeatherStore.response');
 
   @override
-  WeatherResponseModel get response {
+  WeatherResponseModel? get response {
     _$responseAtom.reportRead();
     return super.response;
   }
 
   @override
-  set response(WeatherResponseModel value) {
+  set response(WeatherResponseModel? value) {
     _$responseAtom.reportWrite(value, super.response, () {
       super.response = value;
     });
